requirements.txt
# Bleeding edge Django
django==1.8.7

django-localflavor==1.1
git+https://github.com/ovidner/django-saml-service-provider.git@71bce932726ed932e9616594bfe744e072c6a16a#egg=django_saml_service_provider-master

# REST Framework
djangorestframework==3.3.1
drf-nested-routers==0.11.1
django-rest-auth==0.6.0
dry-rest-permissions==0.1.4

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.6

gunicorn==19.4.1
raven==5.8.1

# Forms
django-braces==1.8.1

# Models
django-model-utils==2.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
# django-allauth==0.23.0

# For the persistence stores
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0

celery==3.1.19

stripe==1.27.1

coverage==4.0.3
flake8==2.5.0
Sphinx
django-extensions==1.5.9
Werkzeug==0.11.2
django-test-plus==1.0.11
factory_boy==2.5.2
PyHamcrest==1.8.5

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4.0

requests==2.8.1
newrelic
